Twentytwoth commit including: add and handle updateTotalPrice, paymentByCard, paymentByCash and setStatus methods to the CartServiceImpl class. Update annotation for updateById method of the CartLineItemController class. Update myCart method of the CartServiceImpl class. Add and handle myCart method to the CartController class. Update CartLineItemDTOResponse method of the CartLineItemDTOResponse class. Update CartDTOResponse method ot the CartDTOResponse class. Add and handle myCartLineItem method to the CartLineItemServiceImpl class. Update useById, releaseById and disableById methods of the UserVoucherServiceImpl class. Add MyCartDTOResponse class to the DTO package. Add and handle MyCartDTOResponse method to the MyCartDTOResponse class. add CartPaymentDTORequest class to the DTO package. Add and handle paymentByCard, paymentByCash and setStatus methods to the CartController class. Update blockById and unblockById methods of the UserController class.

+++ src/main/java/com/r2s/R2Sshop/DTO/VoucherDTORequest.java
@@ -21,7 +21,7 @@ public class VoucherDTORequest {
     private String code;
     @NotNull(message = "Discount cannot be empty !!!")
     @Min(value = 0, message = "Discount must be greater than or equal to 0 !!!")
-    private Integer discount;
+    private Double discount;
     @NotNull(message = "Quantity cannot be empty !!!")
     @Min(value = 1, message = "Quantity must be at least 1 !!!")
     private Integer quantity;
